+++ next.config.mjs
@@ -54,7 +54,6 @@ const withMDX = createMDX({
         {
           themes: {
             light: 'github-light',
-            dark: 'github-dark',
           },
         },
       ],